requirements.txt
Flask>=3.1.1
requests>=2.31.0
fpdf2>=2.7.8
python-dotenv>=1.0.1 